refactoring for sign up api
